Stop using size_t for type sizes/alignments
We want to allow compiling for 64-bit targets from a 32-bit host, so a
type's size may exceed UINT32_MAX even if sizeof(size_t)==4. To
accomodate this, type sizes now use type uint64_t. Alignments use
uint8_t, since they will never exceed 8 (potentially 16 in the future if
we add 128-bit integer types, but still well within the bounds of an
8-byte integer).

Note that a previous commit (2312721) made it an error if a type's size
would overflow. This currently still only works for 64-bit targets,
since it checks if the value literally overflows on the host. This will
need to be updated in the future to check if the value exceeds the
target maximum for `size` if it's smaller than 2**64-1.

+++ include/type_store.h
@@ -1,5 +1,6 @@
 #ifndef HARE_TYPESTORE_H
 #define HARE_TYPESTORE_H
+#include <stdint.h>
 #include "ast.h"
 #include "lex.h"
 #include "types.h"
@@ -14,8 +15,8 @@ struct type_bucket {
 struct context;
 
 struct dimensions {
-	size_t size;
-	size_t align;
+	uint64_t size;
+	uint8_t align;
 };
 
 typedef struct type_bucket *type_store[TYPE_STORE_BUCKETS];
